requirements.txt
torch>=1.0
torchvision>=0.2
numpy>=1.18
pillow>=7.0
tqdm>=4.0
wandb>=0.10.33
pytorch-metric-learning>=1.3.2